i forgot to push

+++ common/src/main/java/it/hurts/octostudios/clavis/common/network/ChunkListeners.java
@@ -23,7 +23,7 @@ public class ChunkListeners {
                 return;
             }
 
-            float difficulty = LootUtils.calculateDifficulty(level, blockPos, randomizable, 0);
+            float difficulty = (float) LootUtils.calculateDifficulty(level, blockPos, randomizable, 0);
             if (difficulty < 0.05f) {
                 return;
             }
